Remove duplicate plugins from the list during each revision fetch.
Previously, if a large number of revisions were being fetched (e.g., 100,000), the number of duplicates would grow exponentially and the process could run out of memory before it purged them at the end of the loop.

Purging them on each iteration is more expensive, but necessary to keep the process alive in those circumstances.

See #9